refactor: extract god class into focused services with proper DI
Extract 4 service classes from the 1204-line Installer to make it a
smaller orchestrator (~775 lines):

- DefinitionResolver: env var definition merging, validation, display
- ConsoleCommandRunner: Symfony console command execution
- BundleInstaller: bundle registration and installation
- PostInstallRunner: post-install command collection and execution

All services use required constructor injection (no nullable/lazy-getter
patterns). The container autowires the full object graph via a catch-all
resource directive in services.yaml. Tests use a shared createInstaller()
factory method from InstallBundleTestHelperTrait.
